Add info regarding zones for satellite
We recently had a beta customer who provisioned three 8x32 data nodes
but they were all provisioned in the same zone...instead of one in
each zone.

Customer mentioned it may be helpful if the doc mentioned the nodes
should be balanced across the zones being deployed to.

So adding the reference to help the customer.