Add info about sat quick start templates and hosts
We had a customer use the Sat quick start templates to provision
a sat location and hosts.

Those templates only allow for one type/size of host, but Event
Streams needs two different types of hosts.

Adding some text to make the reader aware of why we do not want
them to use the templates, an option to partially use it, and
what Event Streams expects during provision.