Add info about Satellite storage auto update
The Satellite Storage team delivered a change that allows for their
function to get auto updated for revision updates.

Telling a customer that is provisioning Event Streams for Sat to
enable this auto update is a good thing...for them and us.

It gets compliance fixes auto applied for us so we don't get
compliance issues, and it helps the customer stay current with
fixes and not require them to take a manually action, periodically.